chore: shrink dead TPC-DI transformations
